+++ nosql_schema/__init__.py
@@ -1,2 +1,3 @@
 from .schema import Schema
 from .db import create_handler
+from .helper import SchemaId