+++ nosql_schema/db/nosqlite/collection_handler.py
@@ -1,4 +1,5 @@
 from ...db import AbstractCollectionHandler
+from ...helper import SchemaId
 
 
 class CollectionHandler(AbstractCollectionHandler):
@@ -80,16 +81,25 @@ class CollectionHandler(AbstractCollectionHandler):
         if not isinstance(query, dict):
             return {}
 
-        for k,v in query.iteritems():
-            if k == '_id':
+        new_query = query.copy()
+
+        for k,v in new_query.iteritems():
+            if isinstance(v, SchemaId):
+                if v.is_list:
+                    new_query[k] = [int(id_) for id_ in v.id_list]
+                else:
+                    new_query[k] = int(v.id_)
+            elif k == '_id':
                 if isinstance(v, dict):
-                    query[k] = CollectionHandler.convert_ids(v, True)
+                    new_query[k] = CollectionHandler.convert_ids(v, True)
                 elif type(v) in [str, unicode]:
-                    query[k] = int(v)
+                    new_query[k] = int(v)
             elif is_id:
                 if type(v) in [str, unicode]:
-                    query[k] = int(v)
+                    new_query[k] = int(v)
                 elif type(v) == list:
-                    query[k] = [int(id_) for id_ in v]
+                    new_query[k] = [int(id_) for id_ in v]
+            elif isinstance(v, dict):
+                new_query[k] = CollectionHandler.convert_ids(v)
 
-        return query
+        return new_query